Programação das cutscenes técnicamente pronta
Sim, eu fiquei de saco cheio de programar o minigame. Tô pensando em trocar o nome da branch, mas tô com medo de que isso dê merda, ent vai ficar assim.

Fiz um prefab, um script e um placeholder para as cutscenes. Ela vai servir para ser utilizada em todos os minigames, só será necessário colocar as animações corretas. A funcionalidade de pular cutscene já existe, mas eu gostaria de mexer um pouco mais nela, só tocar na tela não é satisfatório pra mim, quero que precise pressionar e segurar por 2 segundos.

Próximos passos: Fazer o que acabei de dizer acima, fazer as portas abrirem e revelarem se a resposta está correta ou não antes de passar pra próxima palavra.
