+++ Projeto_Integrador_V/Assets/Placeholders/Cutscene_Placeholder.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7ee5bbb1cb3f82d4488c8f48b719ace6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
