+++ Projeto_Integrador_V/Assets/Placeholders/Cutscene_Placeholder/Cutscene_Video_Placeholder.mp4.meta
@@ -1,0 +1,18 @@
+fileFormatVersion: 2
+guid: 69c8552f2bcbcc84aa640b3dc4fb9c2e
+VideoClipImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  frameRange: 0
+  startFrame: -1
+  endFrame: -1
+  colorSpace: 0
+  deinterlace: 0
+  encodeAlpha: 0
+  flipVertical: 0
+  flipHorizontal: 0
+  importAudio: 1
+  targetSettings: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
