+++ Projeto_Integrador_V/Assets/Prefabs/Prefab_Cutscene_UI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cc2bb3d0ef85fef469f81d5b194c5862
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
