--- Projeto_Integrador_V/Assets/Scripts/Folder_Evite_A_Isca/Game_Manager_Evite_A_Isca.cs
@@ -1,457 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using TMPro;
-using UnityEditor;
-using UnityEngine;
-
-public class Game_Manager_Evite_A_Isca : MonoBehaviour
-{
-    string Right_Word, 
-        Right_Domain;
-
-    public GameObject Door_Prefab;
-
-    //Deixando a lista de Possible_Names para caso a gente abandone a ideia de sílabas e siga por esse lado mais simples.
-    List<string> Possible_Names = new List<string> { "Google", "OLX", "YouTube", "mais um" },
-        Possible_Domains = new List<string> { ".com", ".org", ".gov", ".edu", ".info", ".io", ".net", ".online", ".blog", ".app" },
-        Possible_Progressions = new List<string> { "Doors", "Time", "Difficult", "Syllables" },
-
-        //Eu não sei se esse sistema de sílabas ficou legal. Parece uma ideia legal mas ele fica estranho bem rápido :(
-        Syllables_Consonants_Starters = new List<string> { "b", "c", "d", "f", "g", "h", "j", "k", "l", "m", 
-            "n", "p", "q", "r", "s", "t", "v", "w", "x", "z", 
-            "br", "cr", "dr", "fr", "gr", "kr", "pr", "tr", "vr", 
-            "ch", "lh", "nh", 
-            "bl", "cl", "fl", "gl", "kl", "pl", "tl"},
-        Syllables_Consonants_Finishers = new List<string> { "l", "m", "n", "r", "s", "w", "x", "y", "z" },
-        Syllables_Vowels = new List<string> { "a", "e", "i", "o", "u", "y", 
-            "ae", "ai", "ao", "au", "ay", 
-            "ea", "ei", "eo", "eu", "ey", 
-            "ia", "ie", "io", "iu", "iy", 
-            "oa", "oe", "oi", "ou", "oy", 
-            "ua", "ue", "ui", "uo", "uy", 
-            "ya", "ye", "yi", "yo", "yu"};
-
-    List<string> Generate_All_Words()
-    {
-        List<string> Words = new List<string>();
-
-        for (int i = 0; i < Door_Amount; i++)
-        {
-            Words.Add(Generating_Wrong_Words());
-        }
-
-        return Words;
-    }
-
-    string Generate_Syllable_Word()
-    {
-        string Word = "";
-
-        for (int i = 0; i < Syllables; i++)
-        {
-            string New_Syllable = "";
-
-            
-            if (Random.value > 0.5f)
-            {
-                New_Syllable += Syllables_Consonants_Starters[ Random.Range(0, Syllables_Consonants_Starters.Count) ];
-            }
-
-            
-            New_Syllable += Syllables_Vowels[ Random.Range(0, Syllables_Vowels.Count) ];
-
-            
-            if (Random.value > 0.5f)
-            {
-                New_Syllable += Syllables_Consonants_Finishers[ Random.Range(0, Syllables_Consonants_Finishers.Count) ];
-            }
-
-            Word += New_Syllable;
-        }
-
-        return Word;
-    }
-
-    List<GameObject> Doors_List = new List<GameObject>();
-
-    float Max_Timer, 
-        Current_Timer;
-
-    public int Player_Score, 
-        Player_Highscore;
-
-    [SerializeField] public TextMeshProUGUI Right_Word_Text, 
-        Timer_Display, 
-        Score_Display;
-
-    bool Is_On_Round;
-
-    int Difficult_Level, 
-        Door_Amount,
-        Syllables;
-
-    public float Spacing = 2f;
-
-    
-
-    void Start()
-    {
-        Player_Score = 0; 
-        Door_Amount = 2; 
-        Difficult_Level = 1;
-        Syllables = 1;
-        Max_Timer = 30f;
-
-        Start_New_Round();
-    }
-
-    void Update()
-    {
-
-        Time_Ticking();
-        Touching_Screen();
-    }
-
-    void Generating_Correct_Word()
-    {
-        Right_Word = Generate_Syllable_Word();
-        Right_Domain = (Possible_Domains[Random.Range(0, Possible_Domains.Count)]);
-
-        Right_Word_Text.text = (Right_Word + Right_Domain);
-
-        List<string> s = Generate_All_Words();
-        Spawn_Doors(s);
-    }
-
-    string Generating_Wrong_Words()
-    {
-        string Wrong_Word = Right_Word;
-        string Wrong_Domain = Right_Domain;
-
-        int Randomize_Chances = 0;
-
-
-
-
-        switch (Difficult_Level)
-        {
-            case 1:
-                Randomize_Chances = 0;
-                break;
-
-            case 2:
-                Randomize_Chances = Random.Range(0, 2);
-                break;
-
-            case 3:
-                Randomize_Chances = Random.Range(0, 3);
-                break;
-        }
-
-        if(Randomize_Chances == 0)
-        {
-            Wrong_Domain = Generating_Fake_Domain(Wrong_Domain);
-        }
-        else if (Randomize_Chances == 1)
-        {
-            Randomize_Chances = Random.Range(0, 2);
-
-
-            if (Randomize_Chances == 0)
-            {
-
-                Wrong_Domain = Generating_Fake_Domain(Wrong_Domain);
-            }
-
-            Wrong_Word = Generating_Name_Without_Some_Letter(Wrong_Word);
-        }
-        else if(Randomize_Chances == 2)
-        {
-
-            Randomize_Chances = Random.Range(0, 3);
-
-            if (Randomize_Chances == 0)
-            {
-
-                Wrong_Domain = Generating_Fake_Domain(Wrong_Domain);
-            }
-
-            Wrong_Word = Generating_Fake_Name(Wrong_Word);
-        }
-        
-
-        return Wrong_Word + Wrong_Domain;
-    }
-
-    string Generating_Fake_Domain(string Wrong_Domain)
-    {
-
-
-        string Fake_Domain = Wrong_Domain;
-
-        print(Fake_Domain);
-
-        while (Fake_Domain == Right_Domain)
-        {
-            Fake_Domain = (Possible_Domains[Random.Range(0, Possible_Domains.Count)]);
-        }
-
-        return Fake_Domain;
-    }
-
-    string Generating_Name_Without_Some_Letter(string Wrong_Word)
-    {
-
-        int indexToRemove = Random.Range(0, Wrong_Word.Length);
-
-        string Fake_Name = Wrong_Word.Remove(indexToRemove, 1);
-
-        return Fake_Name;
-
-    }
-
-    string Generating_Fake_Name(string Wrong_Word)
-    {
-        List<string[]> similar_Digits = new List<string[]>
-        {
-        new string[] { "a", "o", "c", "0", "e" },
-        new string[] { "i", "l", "1", "j" },
-        new string[] { "n", "m", "w", "v", "u" },
-        new string[] { "q", "p", "d", "g", "h", "k", "x", "b", "5", "6" },
-        new string[] { "t", "f", "r", "y"},
-        new string[] { "s", "z", "5" }
-        };
-
-        char[] chars = Wrong_Word.ToCharArray();
-
-        
-        int index = Random.Range(0, chars.Length);
-        char original_Char = char.ToLower(chars[index]);
-
-        
-        foreach (var group in similar_Digits)
-        {
-            if (group.Contains(original_Char.ToString()))
-            {
-                
-                string new_Char;
-                do
-                {
-                    new_Char = group[Random.Range(0, group.Length)];
-                }
-                while (new_Char[0] == original_Char);
-
-                chars[index] = new_Char[0];
-                break;
-            }
-        }
-
-        return new string(chars);
-
-
-    }
-
-    void Spawn_Doors(List<string> Names)
-    {
-
-        int correct_Index = Random.Range(0, Door_Amount);
-
-        float total_Width = (Door_Amount - 1) * Spacing;
-        float start_X = -total_Width / 2;
-
-        for(int i = 0; i < Door_Amount; i++)
-        {
-            Vector3 Door_Position = new Vector3(start_X + i * Spacing,
-                Door_Prefab.transform.position.y,
-                Door_Prefab.transform.position.z);
-
-            GameObject Door_Instance = Instantiate(Door_Prefab,
-                Door_Position,
-                Quaternion.identity);
-
-            Doors_List.Add(Door_Instance);
-
-            string textToUse;
-
-            if (i == correct_Index)
-            {
-                textToUse = Right_Word + Right_Domain;
-                Door_Instance.tag = "Correct_Tag_Placeholder";
-            }
-            else
-            {
-                textToUse = Names[i];
-                Door_Instance.tag = "Wrong_Tag_Placeholder";
-            }
-
-
-            TextMeshPro Door_Text = Door_Instance.GetComponentInChildren<TextMeshPro>();
-            Door_Text.text = textToUse;
-        }
-    }
-
-    void Start_New_Round()
-    {
-
-        Score_Display.text = Player_Score.ToString();
-
-        foreach (GameObject door in Doors_List)
-        {
-            Destroy(door);
-        }
-
-        Doors_List.Clear();
-        Current_Timer = Max_Timer;
-        Is_On_Round = true;
-
-        Generating_Correct_Word();
-    }
-
-    void Time_Ticking()
-    {
-        if (Is_On_Round)
-        {
-            Current_Timer -= Time.deltaTime;
-
-            Timer_Display.text = ((int)Current_Timer).ToString();
-
-
-            if (Current_Timer <= 0f)
-            {
-                Defeat();
-            }
-        }
-    }
-
-    void Touching_Screen()
-    {
-        if (Is_On_Round)
-        {
-            if(Input.touchCount > 0)
-            {
-                Touch Getting_Touch = Input.GetTouch(0);
-
-                if(Getting_Touch.phase == TouchPhase.Began)
-                {
-
-                    //Eu não pensei em nomes condinzentes para por nessas variaveis...
-                    Ray r = Camera.main.ScreenPointToRay(Getting_Touch.position);
-                    RaycastHit hit;
-
-                    if(Physics.Raycast(r, out hit))
-                    {
-                        if (hit.transform.root.CompareTag("Correct_Tag_Placeholder"))
-                        {
-                            Right_Ansher();
-                        }
-                        else if (hit.transform.root.CompareTag("Wrong_Tag_Placeholder"))
-                        {
-                            Defeat();
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    void Defeat()
-    {
-
-        print("Você Perdeu!");
-        Is_On_Round = false;
-    }
-
-    void Right_Ansher()
-    {
-
-        Player_Score++;
-        
-
-        if (Player_Score % 5 == 0 && Possible_Progressions.Count != 0)
-        {
-            print("Dificultando");
-            Difficulting();
-        }
-        else if(Possible_Progressions.Count == 0)
-        {
-            print("Não dá pra dificultar mais");
-        }
-
-        Start_New_Round();
-    }
-
-    void Difficulting()
-    {
-
-        //Nome extremamente goofy pra uma variavel, eu sei. Ainda vou achar um nome melhor.
-        string Decide_Whats_Bcome_Harder = (Possible_Progressions[Random.Range(0, Possible_Progressions.Count)]);
-
-        print($"Aumentar dificuldade: {Decide_Whats_Bcome_Harder}");
-
-        switch (Decide_Whats_Bcome_Harder)
-        {
-            case "Doors": Add_Doors(); break;
-            case "Time":    Reduce_Max_Timer(); break;
-            case "Difficult":   Deixando_Mais_Dificil();    break;
-            case "Syllables":   Adding_Syllables(); break;
-
-
-        }
-    }
-
-    void Add_Doors()
-    {
-        Door_Amount++;
-
-        print($"Quantidade de portas: {Door_Amount}");
-
-        if(Door_Amount == 5)
-        {
-
-            Possible_Progressions.Remove("Doors");
-            print("Nível máximo alcançado em PORTAS");
-        }
-    }
-
-    void Reduce_Max_Timer()
-    {
-        Max_Timer -= 5f;
-
-        print($"Tempo máximo: {Max_Timer}");
-
-        if (Max_Timer == 5f)
-        {
-            Possible_Progressions.Remove("Time");
-            print("Nível máximo alcançado em TEMPO");
-        }
-    }
-
-    //Ainda vou trocar o nome desse método. Eu também não gosto de como tem 2 coisas falando sobre "dificultar". Queria um nome diferente
-    void Deixando_Mais_Dificil()
-    {
-        Difficult_Level++;
-
-        print($"Dificuldade nível: {Difficult_Level}");
-
-        if (Difficult_Level == 3)
-        {
-            Possible_Progressions.Remove("Difficult");
-            print("Nível máximo alcançado em DIFICULDADE");
-        }
-    }
-
-    void Adding_Syllables()
-    {
-        Syllables++;
-
-        print($"Quantidade de silabas: {Syllables}");
-
-        if(Syllables == 5)
-        {
-            Possible_Progressions.Remove("Syllables");
-            print("Nível máximo alcançado em SILABAS");
-        }
-    }
-}
