--- Projeto_Integrador_V/Assets/Scripts/Folder_Evite_A_Isca/Game_Manager_Evite_A_Isca.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 94b27392b7390534cba22087a66a699b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
