+++ Projeto_Integrador_V/Assets/Scripts/Folder_Minigames.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 253062c3f015fef4bb0b70a786a7e15f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
