+++ Projeto_Integrador_V/Assets/Scripts/Folder_Minigames/Script_Cutscene_Controller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9235329f93c78c94582ada6d0722ff99
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
